Refactor codebase in preparation for hide PR
