+++ Source/DafnyCore/Verifier/BoogieGenerator.LetExpr.cs
@@ -32,6 +32,113 @@ using static Microsoft.Dafny.GenericErrors;
 namespace Microsoft.Dafny {
   public partial class BoogieGenerator {
 
+    partial class ExpressionTranslator {
+
+
+      private Expr LetCanCallAssumption(LetExpr letExpr) {
+        if (!letExpr.Exact) {
+          // CanCall[[ var b0,b1 :| RHS(b0,b1,g); Body(b0,b1,g,h) ]] =
+          //   $let$canCall(g) &&
+          //   CanCall[[ Body($let$b0(g), $let$b1(g), h) ]]
+          BoogieGenerator.LetDesugaring(letExpr);  // call LetDesugaring to prepare the desugaring and populate letSuchThatExprInfo with something for e
+          var info = BoogieGenerator.letSuchThatExprInfo[letExpr];
+          // $let$canCall(g)
+          var canCall = info.CanCallFunctionCall(BoogieGenerator, this);
+          Dictionary<IVariable, Expression> substMap = new Dictionary<IVariable, Expression>();
+          foreach (var bv in letExpr.BoundVars) {
+            // create a call to $let$x(g)
+            var args = info.SkolemFunctionArgs(bv, BoogieGenerator, this);
+            var call = new BoogieFunctionCall(bv.tok, info.SkolemFunctionName(bv), info.UsesHeap, info.UsesOldHeap, info.UsesHeapAt, args.Item1, args.Item2);
+            call.Type = bv.Type;
+            substMap.Add(bv, call);
+          }
+          var p = Substitute(letExpr.Body, null, substMap);
+          var cc = BplAnd(canCall, CanCallAssumption(p));
+          return cc;
+        } else {
+          // CanCall[[ var b := RHS(g); Body(b,g,h) ]] =
+          //   CanCall[[ RHS(g) ]] &&
+          //   (var lhs0,lhs1,... := rhs0,rhs1,...;  CanCall[[ Body ]])
+          Boogie.Expr canCallRHS = Boogie.Expr.True;
+          foreach (var rhs in letExpr.RHSs) {
+            canCallRHS = BplAnd(canCallRHS, CanCallAssumption(rhs));
+          }
+
+          var bodyCanCall = CanCallAssumption(letExpr.Body);
+          // We'd like to compute the free variables if "bodyCanCall". It would be nice to use the Boogie
+          // routine Bpl.Expr.ComputeFreeVariables for this purpose. However, calling it requires the Boogie
+          // expression to be resolved. Instead, we do the cheesy thing of computing the set of names of
+          // free variables in "bodyCanCall".
+          var vis = new VariableNameVisitor();
+          vis.Visit(bodyCanCall);
+
+          List<Boogie.Variable> lhssAll;
+          List<Boogie.Expr> rhssAll;
+          TrLetExprPieces(letExpr, out lhssAll, out rhssAll);
+          Contract.Assert(lhssAll.Count == rhssAll.Count);
+
+          // prune lhss,rhss to contain only those pairs where the LHS is used in the body
+          var lhssPruned = new List<Boogie.Variable>();
+          var rhssPruned = new List<Boogie.Expr>();
+          for (var i = 0; i < lhssAll.Count; i++) {
+            var bv = lhssAll[i];
+            if (vis.Names.Contains(bv.Name)) {
+              lhssPruned.Add(bv);
+              rhssPruned.Add(rhssAll[i]);
+            }
+          }
+          Boogie.Expr let = lhssPruned.Count == 0 ? bodyCanCall : new Boogie.LetExpr(letExpr.tok, lhssPruned, rhssPruned, null, bodyCanCall);
+          return BplAnd(canCallRHS, let);
+        }
+      }
+
+      private Expr TrLetExpr(LetExpr letExpr) {
+        var e = letExpr;
+        if (!e.Exact) {
+          var d = BoogieGenerator.LetDesugaring(e);
+          return TrExpr(d);
+        } else {
+          TrLetExprPieces(e, out var lhss, out var rhss);
+          // in the translation of body, treat a let-bound variable as IsLit if its RHS definition is IsLit
+          Contract.Assert(lhss.Count == rhss.Count);  // this is a postcondition of TrLetExprPieces
+          var previousCount = BoogieGenerator.letBoundVariablesWithLitRHS.Count;
+          for (var i = 0; i < lhss.Count; i++) {
+            if (BoogieGenerator.IsLit(rhss[i])) {
+              BoogieGenerator.letBoundVariablesWithLitRHS.Add(lhss[i].Name);
+            }
+            i++;
+          }
+          var body = TrExpr(e.Body);
+          foreach (var v in lhss) {
+            BoogieGenerator.letBoundVariablesWithLitRHS.Remove(v.Name);
+          }
+          Contract.Assert(previousCount == BoogieGenerator.letBoundVariablesWithLitRHS.Count);
+          // in the following, use the token for Body instead of the token for the whole let expression; this gives better error locations
+          return new Boogie.LetExpr(GetToken(e.Body), lhss, rhss, null, body);
+        }
+      }
+
+
+      public void TrLetExprPieces(LetExpr let, out List<Boogie.Variable> lhss, out List<Boogie.Expr> rhss) {
+        Contract.Requires(let != null);
+        var substMap = new Dictionary<IVariable, Expression>();
+        for (int i = 0; i < let.LHSs.Count; i++) {
+          var rhs = TrExpr(let.RHSs[i]);
+          var toType = let.LHSs[i].Var?.Type ?? let.LHSs[i].Expr.Type;
+          rhs = BoogieGenerator.CondApplyBox(rhs.tok, rhs, let.RHSs[i].Type, toType);
+          BoogieGenerator.AddCasePatternVarSubstitutions(let.LHSs[i], rhs, substMap);
+        }
+        lhss = new List<Boogie.Variable>();
+        rhss = new List<Boogie.Expr>();
+        foreach (var v in let.BoundVars) {
+          var rhs = substMap[v];  // this should succeed (that is, "v" is in "substMap"), because the AddCasePatternVarSubstitutions calls above should have added a mapping for each bound variable in let.BoundVars
+          var bv = BplBoundVar(v.AssignUniqueName(BoogieGenerator.currentDeclaration.IdGenerator), BoogieGenerator.TrType(v.Type), out var bvIde);
+          lhss.Add(bv);
+          rhss.Add(TrExpr(rhs));
+        }
+      }
+    }
+
     /// <summary>
     /// Fills in, if necessary, the e.translationDesugaring field, and returns it.
     /// Also, makes sure that letSuchThatExprInfo maps e to something.
